+++ toolkit/block-production-log/primitives/src/lib.rs
@@ -9,19 +9,26 @@
 //!
 //! ### Adding to the node
 //!
+//! #### Defining the types
+//!
+//! The feature requires two types to be defind by the chain builders in their code:
+//! - `Author`: the type representing the block author, as it's configured in the feature's pallet
+//! - `Moment`: a moment in time when the block was produced, which carries enough information to
+//!             calculate the block's author. Typcally, this type can be a timestamp, or a slot,
+//!             depending on the consensus mechanism used, but can be a richer type if needed.
+//!
 //! #### Implementing the runtime API
 //!
 //! The block production log feature requires [BlockProductionLogApi] to be implemented by the Partner Chain runtime
 //! so the current block author can be identified. The concrete implementation must use or match the mechanism of
-//! block author selection used by the particular Partner Chain's consensus mechanism. The API should use the same
-//! type `CommitteeMember` to represent the block author as was configured in pallet's configuration.
+//! block author selection used by the particular Partner Chain's consensus mechanism.
 //!
 //! An example for a Partner Chain using Aura consensus looks like this:
 //! ```rust, ignore
 //! impl_runtime_apis! {
-//!     impl BlockProductionLogApi<Block, CommitteeMember> for Runtime {
+//!     impl BlockProductionLogApi<Block, CommitteeMember, Slot> for Runtime {
 //!         fn get_author(slot: Slot) -> Option<CommitteeMember> {
-//!             SessionCommitteeManagement::get_current_authority_round_robin(*slot as usize)
+//!             SessionCommitteeManagement::get_current_authority_round_robin(u64::from(*slot) as usize)
 //!         }
 //!     }
 //! }
@@ -41,6 +48,8 @@
 //! // Return the inherent data provider together with other IDPs
 //! Ok((timestamp_idp, slot_idp, ..., block_author_idp, ...))
 //! ```
+//!
+//! Notice that the IDP must be passed the current `Moment`, which in this instance is `Slot`.
 //!
 //! The inherent data provider created using `BlockAuthorInherentProvider::new` will check whether `BlockProductionLogApi`
 //! is available in the runtime and will only provide inherent data if the API is present.
@@ -87,55 +96,68 @@ impl IsFatalError for InherentError {
 	}
 }
 
+/// Current block producer data used as inherent data
+#[derive(Clone, Debug, Encode, Decode, PartialEq, Eq)]
+pub struct BlockProductionInherentDataV1<Moment, BlockProducerId> {
+	/// Moment of the block
+	pub moment: Moment,
+	/// Block producer ID
+	pub block_producer_id: BlockProducerId,
+}
+
 /// Inherent data provider providing the block author of the current block
 /// Type parameters:
 /// - `Author`: Type representing a block author.
 #[cfg(feature = "std")]
 #[derive(Debug)]
-pub struct BlockAuthorInherentProvider<Author> {
+pub struct BlockAuthorInherentProvider<Moment, Author> {
 	/// Optional value of the current block author. Inherent data is not provided if [None].
-	pub author: Option<Author>,
+	pub data: Option<BlockProductionInherentDataV1<Moment, Author>>,
 }
 
 #[cfg(feature = "std")]
-impl<Author> BlockAuthorInherentProvider<Author> {
+impl<Moment, Author> BlockAuthorInherentProvider<Moment, Author> {
 	/// Creates a new [BlockAuthorInherentProvider] using runtime API [BlockProductionLogApi].
 	///
 	/// The inherent data provider returned will be inert if [BlockProductionLogApi] is not detected in the runtime.
-	pub fn new<C, Member, Block>(
+	pub fn new<C, Block>(
 		client: &C,
 		parent_hash: Block::Hash,
-		slot: sidechain_slots::Slot,
+		moment: Moment,
 	) -> Result<Self, Box<dyn core::error::Error + Send + Sync>>
 	where
-		Member: Decode,
+		Moment: Encode,
+		Author: Decode,
 		Block: BlockT,
 		C: ProvideRuntimeApi<Block>,
-		C::Api: BlockProductionLogApi<Block, Member>,
-		Author: From<Member>,
+		C::Api: BlockProductionLogApi<Block, Author, Moment>,
 	{
 		let api = client.runtime_api();
-		if !api.has_api::<dyn BlockProductionLogApi<Block, Member>>(parent_hash)? {
-			return Ok(Self { author: None });
+		if !api.has_api::<dyn BlockProductionLogApi<Block, Author, Moment>>(parent_hash)? {
+			return Ok(Self { data: None });
 		}
-		let author = client.runtime_api().get_author(parent_hash, slot)?.map(Author::from);
+		let data = client
+			.runtime_api()
+			.get_author(parent_hash, &moment)?
+			.map(|author| BlockProductionInherentDataV1 { moment, block_producer_id: author });
 
-		Ok(BlockAuthorInherentProvider { author })
+		Ok(BlockAuthorInherentProvider { data })
 	}
 }
 
 #[cfg(feature = "std")]
 #[async_trait::async_trait]
-impl<T> sp_inherents::InherentDataProvider for BlockAuthorInherentProvider<T>
+impl<Moment, T> sp_inherents::InherentDataProvider for BlockAuthorInherentProvider<Moment, T>
 where
 	T: Send + Sync + Encode + Decode,
+	Moment: Send + Sync + Encode + Decode,
 {
 	async fn provide_inherent_data(
 		&self,
 		inherent_data: &mut InherentData,
 	) -> Result<(), sp_inherents::Error> {
-		if let Some(author) = &self.author {
-			inherent_data.put_data(INHERENT_IDENTIFIER, author)?;
+		if let Some(data) = &self.data {
+			inherent_data.put_data(INHERENT_IDENTIFIER, data)?;
 		}
 		Ok(())
 	}
@@ -157,11 +179,12 @@ where
 sp_api::decl_runtime_apis! {
 	/// Runtime API exposing data required for the [BlockAuthorInherentProvider] to operate.
 	/// Type parameters:
-	/// - `Member`: type representing a committee member eligible to be a block author. This type should correspond
+	/// - `Author`: type representing a committee member eligible to be a block author. This type should correspond
 	///             to what is configured as the block author type used by the pallet.
-	pub trait BlockProductionLogApi<Member>
+	pub trait BlockProductionLogApi<Author, Moment>
 	where
-		Member: Decode
+		Author: Decode,
+		Moment: Encode
 	{
 		/// Function returning the current block's author.
 		///
@@ -169,7 +192,9 @@ sp_api::decl_runtime_apis! {
 		/// independently calculate it, or obtain it from another source.
 		///
 		/// Parameters:
-		/// - `slot`: slot number of the block currently being produced
-		fn get_author(slot: sidechain_slots::Slot) -> Option<Member>;
+		/// - `moment`: value representing the current block's moment in time. It will typically be a timestamp or
+		///             slot but can be any other value that is needed to decide the current block's author and can
+		///             be passed from the node.
+		fn get_author(moment: &Moment) -> Option<Author>;
 	}
 }